move enums and interfaces into their own files for better readability
