+++ src/architecture/layer/GRULayer.ts
@@ -1,8 +1,11 @@
 import {Layer} from "./Layer";
-import {Activation, ActivationType} from "../../methods/Activation";
-import {Node, NodeType} from "../Node";
-import {ConnectionType, GatingType} from "../Architect";
+import {Activation} from "../../methods/Activation";
+import {Node} from "../Node";
 import {Connection} from "../Connection";
+import {ActivationType} from "../../enums/ActivationType";
+import {NodeType} from "../../enums/NodeType";
+import {ConnectionType} from "../../enums/ConnectionType";
+import {GatingType} from "../../enums/GatingType";
 
 export class GRULayer extends Layer {
     constructor(outputSize: number, options: { activationType?: ActivationType } = {}) {
