+++ src/architecture/layer/Layer.ts
@@ -1,6 +1,7 @@
 import {Connection} from "../Connection";
 import {Node} from "../Node";
-import {ConnectionType, GatingType} from "../Architect";
+import {ConnectionType} from "../../enums/ConnectionType";
+import {GatingType} from "../../enums/GatingType";
 
 export abstract class Layer {
     public outputSize: number;
