+++ src/architecture/layer/OutputLayer.ts
@@ -1,5 +1,6 @@
 import {Layer} from "./Layer";
-import {Node, NodeType} from "../Node";
+import {Node} from "../Node";
+import {NodeType} from "../../enums/NodeType";
 
 export class OutputLayer extends Layer {
     constructor(outputSize: number) {
