+++ src/enums/ActivationType.ts
@@ -1,0 +1,18 @@
+export enum ActivationType {
+    NO_ACTIVATION,
+    LogisticActivation,
+    TanhActivation,
+    IdentityActivation,
+    StepActivation,
+    RELUActivation,
+    SoftSignActivation,
+    SinusoidActivation,
+    GaussianActivation,
+    BentIdentityActivation,
+    BipolarActivation,
+    BipolarSigmoidActivation,
+    HardTanhActivation,
+    AbsoluteActivation,
+    InverseActivation,
+    SELUActivation
+}
