+++ src/enums/ConnectionType.ts
@@ -1,0 +1,4 @@
+export enum ConnectionType {
+    ALL_TO_ALL,
+    ONE_TO_ONE
+}
