+++ src/enums/GatingType.ts
@@ -1,0 +1,3 @@
+export enum GatingType {
+    INPUT, SELF, OUTPUT
+}
