+++ src/enums/NodeType.ts
@@ -1,0 +1,3 @@
+export enum NodeType {
+    INPUT, HIDDEN, OUTPUT
+}
