+++ src/interfaces/ConnectionJSON.ts
@@ -1,0 +1,6 @@
+export interface ConnectionJSON {
+    weight: number;
+    fromIndex: number;
+    toIndex: number;
+    gateNodeIndex: number | null;
+}
